requirements.txt
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
qrcode==8.0
sqlparse==0.5.3
tzdata==2025.2
Werkzeug==3.1.3
